SoftwareDocumentCode/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Libraries
torch==2.1.0
torchvision==0.16.0
ultralytics==8.1.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
scikit-learn==1.3.2

# Computer Vision
opencv-contrib-python==4.8.1.78
imageio==2.31.6
imageio-ffmpeg==0.4.9

# Geospatial
geopy==2.4.1
folium==0.15.1
geopandas==0.14.1
shapely==2.0.2

# Data Processing
pandas==2.1.3
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Report Generation
reportlab==4.0.7
fpdf2==2.7.6
jinja2==3.1.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
supabase==2.3.0
postgrest==0.13.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

